+++ src/main/java/com/capgemini/cn/demo/userSystem/service/BranchInfoService.java
@@ -14,7 +14,7 @@ public interface BranchInfoService {
 	void updateBranchInfo(BranchInfo branchInfo);
 	void deleteBranchInfoById(IdToBeJson branchId);
 	
-	int selectBranchInfoIdByName(String name);
+	long selectBranchInfoIdByName(String name);
 
 
 }
